+++ chains/container-chains/nodes/frontier/src/command.rs
@@ -17,7 +17,7 @@
 use {
     crate::{
         chain_spec,
-        cli::{Cli, RelayChainCli, Subcommand},
+        cli::{Cli, Subcommand},
         service::{self, frontier_database_dir, NodeConfig},
     },
     container_chain_template_frontier_runtime::Block,
@@ -29,15 +29,14 @@ use {
     dc_orchestrator_chain_interface::OrchestratorChainInterface,
     frame_benchmarking_cli::{BenchmarkCmd, SUBSTRATE_REFERENCE_HARDWARE},
     log::{info, warn},
-    node_common::{command::generate_genesis_block, service::NodeBuilderConfig as _},
+    node_common::{command::generate_genesis_block, service::NodeBuilderConfig as _, cli::RelayChainCli, chain_spec as node_common_chain_spec},
     parity_scale_codec::Encode,
     polkadot_cli::IdentifyVariant,
     sc_cli::{
-        ChainSpec, CliConfiguration, DefaultConfigurationValues, ImportParams, KeystoreParams,
-        NetworkParams, Result, SharedParams, SubstrateCli,
+        ChainSpec,
+        Result, SubstrateCli,
     },
     sc_service::{
-        config::{BasePath, PrometheusConfig},
         DatabaseSource, KeystoreContainer, TaskManager,
     },
     sc_telemetry::TelemetryWorker,
@@ -319,7 +318,7 @@ pub fn run() -> Result<()> {
                         sc_sysinfo::gather_hwbench(Some(database_path), &SUBSTRATE_REFERENCE_HARDWARE)
                     })).flatten();
 
-                let para_id = chain_spec::Extensions::try_get(&*config.chain_spec)
+                let para_id = node_common_chain_spec::Extensions::try_get(&*config.chain_spec)
                     .map(|e| e.para_id)
                     .ok_or("Could not find parachain ID in chain-spec.")?;
 
@@ -335,7 +334,7 @@ pub fn run() -> Result<()> {
                     max_past_logs: cli.run.max_past_logs,
                 };
 
-                let extension = chain_spec::Extensions::try_get(&*config.chain_spec);
+                let extension = node_common_chain_spec::Extensions::try_get(&*config.chain_spec);
 
                 let relay_chain_id = extension.map(|e| e.relay_chain.clone());
 
@@ -397,131 +396,6 @@ pub fn run() -> Result<()> {
     }
 }
 
-impl DefaultConfigurationValues for RelayChainCli {
-    fn p2p_listen_port() -> u16 {
-        30334
-    }
-
-    fn rpc_listen_port() -> u16 {
-        9945
-    }
-
-    fn prometheus_listen_port() -> u16 {
-        9616
-    }
-}
-
-impl CliConfiguration<Self> for RelayChainCli {
-    fn shared_params(&self) -> &SharedParams {
-        self.base.base.shared_params()
-    }
-
-    fn import_params(&self) -> Option<&ImportParams> {
-        self.base.base.import_params()
-    }
-
-    fn network_params(&self) -> Option<&NetworkParams> {
-        self.base.base.network_params()
-    }
-
-    fn keystore_params(&self) -> Option<&KeystoreParams> {
-        self.base.base.keystore_params()
-    }
-
-    fn base_path(&self) -> Result<Option<BasePath>> {
-        Ok(self
-            .shared_params()
-            .base_path()?
-            .or_else(|| Some(self.base_path.clone().into())))
-    }
-
-    fn rpc_addr(&self, default_listen_port: u16) -> Result<Option<Vec<sc_cli::RpcEndpoint>>> {
-        self.base.base.rpc_addr(default_listen_port)
-    }
-
-    fn prometheus_config(
-        &self,
-        default_listen_port: u16,
-        chain_spec: &Box<dyn ChainSpec>,
-    ) -> Result<Option<PrometheusConfig>> {
-        self.base
-            .base
-            .prometheus_config(default_listen_port, chain_spec)
-    }
-
-    fn init<F>(&self, _support_url: &String, _impl_version: &String, _logger_hook: F) -> Result<()>
-    where
-        F: FnOnce(&mut sc_cli::LoggerBuilder),
-    {
-        unreachable!("PolkadotCli is never initialized; qed");
-    }
-
-    fn chain_id(&self, is_dev: bool) -> Result<String> {
-        let chain_id = self.base.base.chain_id(is_dev)?;
-
-        Ok(if chain_id.is_empty() {
-            self.chain_id.clone().unwrap_or_default()
-        } else {
-            chain_id
-        })
-    }
-
-    fn role(&self, is_dev: bool) -> Result<sc_service::Role> {
-        self.base.base.role(is_dev)
-    }
-
-    fn transaction_pool(&self, is_dev: bool) -> Result<sc_service::config::TransactionPoolOptions> {
-        self.base.base.transaction_pool(is_dev)
-    }
-
-    fn trie_cache_maximum_size(&self) -> Result<Option<usize>> {
-        self.base.base.trie_cache_maximum_size()
-    }
-
-    fn rpc_methods(&self) -> Result<sc_service::config::RpcMethods> {
-        self.base.base.rpc_methods()
-    }
-
-    fn rpc_max_connections(&self) -> Result<u32> {
-        self.base.base.rpc_max_connections()
-    }
-
-    fn rpc_cors(&self, is_dev: bool) -> Result<Option<Vec<String>>> {
-        self.base.base.rpc_cors(is_dev)
-    }
-
-    fn default_heap_pages(&self) -> Result<Option<u64>> {
-        self.base.base.default_heap_pages()
-    }
-
-    fn force_authoring(&self) -> Result<bool> {
-        self.base.base.force_authoring()
-    }
-
-    fn disable_grandpa(&self) -> Result<bool> {
-        self.base.base.disable_grandpa()
-    }
-
-    fn max_runtime_instances(&self) -> Result<Option<usize>> {
-        self.base.base.max_runtime_instances()
-    }
-
-    fn announce_block(&self) -> Result<bool> {
-        self.base.base.announce_block()
-    }
-
-    fn telemetry_endpoints(
-        &self,
-        chain_spec: &Box<dyn ChainSpec>,
-    ) -> Result<Option<sc_telemetry::TelemetryEndpoints>> {
-        self.base.base.telemetry_endpoints(chain_spec)
-    }
-
-    fn node_name(&self) -> Result<String> {
-        self.base.base.node_name()
-    }
-}
-
 fn rpc_provider_mode(cli: Cli, profile_id: u64) -> Result<()> {
     log::info!("Starting in RPC provider mode!");
 
@@ -575,7 +449,7 @@ fn rpc_provider_mode(cli: Cli, profile_id: u64) -> Result<()> {
 
             log::info!("Container chain CLI: {container_chain_cli:?}");
 
-            let para_id = chain_spec::Extensions::try_get(&*config.chain_spec)
+            let para_id = node_common_chain_spec::Extensions::try_get(&*config.chain_spec)
                 .map(|e| e.para_id)
                 .ok_or("Could not find parachain ID in chain-spec.")?;
 
@@ -623,7 +497,7 @@ fn rpc_provider_mode(cli: Cli, profile_id: u64) -> Result<()> {
                 .await
                 .map_err(|e| sc_service::Error::Application(Box::new(e) as Box<_>))?;
 
-            let relay_chain = crate::chain_spec::Extensions::try_get(&*config.chain_spec)
+            let relay_chain = node_common_chain_spec::Extensions::try_get(&*config.chain_spec)
                 .map(|e| e.relay_chain.clone())
                 .ok_or("Could not find relay_chain extension in chain-spec.")?;
